a testing-ready boilerplate
remove unnecessary files
