+++ package.json
@@ -25,12 +25,8 @@
     "mock",
     "mocking"
   ],
-  "author": "Matt Zabriskie",
+  "author": "Matt Zabriskie, Mona Zhuang",
   "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/mzabriskie/moxios/issues"
-  },
-  "homepage": "https://github.com/mzabriskie/moxios#readme",
   "peerDependencies": {
     "axios": ">= 0.13.0"
   },
@@ -51,6 +47,7 @@
     "karma-webpack": "^1.7.0",
     "lolex": "^1.5.0",
     "mocha": "^2.5.3",
+    "moxios": "^0.4.0",
     "sinon": "^1.17.4",
     "webpack": "^1.13.1",
     "webpack-dev-server": "^1.14.1"
